requirements.txt
django
djangorestframework
drf-spectacular
django-cors-headers
gunicorn
pytz
tzdata
psycopg2-binary
python-dotenv
requests
